refactor: Moved RequestQueue to utils
